nptl: x32: Fix Race conditions in pthread cancellation [BZ#12683]
This patches adds the x32 modification required for the BZ#12683.
It follows the x86_64-x32 ABI and pointers are zero-extended.
However, compiler may not see such cases and accuse a cast from pointer
to integer of different size and for such cases the warning is
explict disabled.

Checked on x86_64-linux-gnu-x32.